+++ src/main/ts/components/Editor.tsx
@@ -1,44 +1,141 @@
 import * as React from 'react';
+import type { Bookmark, EditorEvent, TinyMCE, Editor as TinyMCEEditor } from 'tinymce';
 import { IEvents } from '../Events';
 import { ScriptItem, ScriptLoader } from '../ScriptLoader2';
 import { getTinymce } from '../TinyMCE';
-import { isFunction, isTextareaOrInput, mergePlugins, uuid, configHandlers, isBeforeInputEventAvailable, isInDoc, setMode } from '../Utils';
+import { configHandlers, isBeforeInputEventAvailable, isFunction, isInDoc, isTextareaOrInput, mergePlugins, setMode, uuid } from '../Utils';
 import { EditorPropTypes, IEditorPropTypes } from './EditorPropTypes';
-import { Bookmark, Editor as TinyMCEEditor, EditorEvent, TinyMCE } from 'tinymce';
+
+const changeEvents = 'change keyup compositionend setcontent CommentChange';
+
+type OmitStringIndexSignature<T> = { [K in keyof T as string extends K ? never : K]: T[K] };
+
+interface DoNotUse<T extends string> {
+  __brand: T;
+}
+
+type OmittedInitProps = 'selector' | 'target' | 'readonly' | 'license_key';
 
 type EditorOptions = Parameters<TinyMCE['init']>[0];
+
+export type InitOptions = Omit<OmitStringIndexSignature<EditorOptions>, OmittedInitProps> & {
+  selector?: DoNotUse<'selector prop is handled internally by the component'>;
+  target?: DoNotUse<'target prop is handled internally by the component'>;
+  readonly?: DoNotUse<'readonly prop is overridden by the component, use the `disabled` prop instead'>;
+  license_key?: DoNotUse<'license_key prop is overridden by the integration, use the `licenseKey` prop instead'>;
+} & { [key: string]: unknown };
 
 export type Version = `${'4' | '5' | '6' | '7'}${'' | '-dev' | '-testing' | `.${number}` | `.${number}.${number}`}`;
 
 export interface IProps {
+  /**
+   * @see {@link https://www.tiny.cloud/docs/tinymce/7/react-ref/#apikey React Tech Ref - apiKey}
+   * @description TinyMCE API key for deployments using Tiny Cloud.
+   */
   apiKey: string;
+  /**
+   * @see {@link https://www.tiny.cloud/docs/tinymce/7/react-ref/#id React Tech Ref - id}
+   * @description The ID of the element to render the editor into.
+   */
   id: string;
+  /**
+   * @see {@link https://www.tiny.cloud/docs/tinymce/7/react-ref/#inline React Tech Ref - inline}
+   * @description Whether the editor should be rendered inline. Equivalent to the `inline` option in TinyMCE.
+   */
   inline: boolean;
+  /**
+   * @see {@link https://www.tiny.cloud/docs/tinymce/7/react-ref/#initialvalue React Tech Ref - initialValue}
+   * @description The initial HTML content of the editor.
+   *
+   * IMPORTANT: Ensure that this is **not** updated by `onEditorChange` or the editor will be unusable.
+   */
   initialValue: string;
+  /**
+   * @see {@link https://www.tiny.cloud/docs/tinymce/7/react-ref/#oneditorchange React Tech Ref - onEditorChange}
+   * @description Used to store the state of the editor outside the component. Typically used for controlled components.
+   * @param a The current HTML content of the editor.
+   * @param editor The TinyMCE editor instance.
+   * @returns void
+   */
   onEditorChange: (a: string, editor: TinyMCEEditor) => void;
+  /**
+   * @see {@link https://www.tiny.cloud/docs/tinymce/7/react-ref/#value React Tech Ref - value}
+   * @description The current HTML content of the editor. Typically used for controlled components.
+   */
   value: string;
-  init: EditorOptions & Partial<Record<'selector' | 'target' | 'readonly' | 'license_key', undefined>>;
+  /**
+   * @see {@link https://www.tiny.cloud/docs/tinymce/7/react-ref/#init React Tech Ref - init}
+   * @description Additional settings passed to `tinymce.init()` when initializing the editor.
+   */
+  init: InitOptions;
+  /**
+   * @see {@link https://www.tiny.cloud/docs/tinymce/7/react-ref/#tagname React Tech Ref - tagName}
+   * @description The tag name of the element to render the editor into. Only valid when `inline` is `true`.
+   */
   tagName: string;
+  /**
+   * @see {@link https://www.tiny.cloud/docs/tinymce/7/react-ref/#tabIndex React Tech Ref - tabIndex}
+   * @description The tab index of the element that the editor wraps.
+   */
   tabIndex: number;
+  /**
+   * @see {@link https://www.tiny.cloud/docs/tinymce/7/react-ref/#cloudchannel React Tech Ref - cloudChannel}
+   * @description The TinyMCE build to use when loading from Tiny Cloud.
+   */
   cloudChannel: Version;
+  /**
+   * @see {@link https://www.tiny.cloud/docs/tinymce/7/react-ref/#plugins React Tech Ref - plugins}
+   * @description The plugins to load into the editor. Equivalent to the `plugins` option in TinyMCE.
+   */
   plugins: NonNullable<EditorOptions['plugins']>;
+  /**
+   * @see {@link https://www.tiny.cloud/docs/tinymce/7/react-ref/#toolbar React Tech Ref - toolbar}
+   * @description The toolbar to load into the editor. Equivalent to the `toolbar` option in TinyMCE.
+   */
   toolbar: NonNullable<EditorOptions['toolbar']>;
+  /**
+   * @see {@link https://www.tiny.cloud/docs/tinymce/7/react-ref/#disabled React Tech Ref - disabled}
+   * @description Whether the editor should be "disabled" (read-only).
+   */
   disabled: boolean;
+  /**
+   * @see {@link https://www.tiny.cloud/docs/tinymce/7/react-ref/#textareaname React Tech Ref - textareaName}
+   * @description Set the `name` attribute of the `textarea` element used for the editor in forms. Only valid in iframe mode.
+   */
   textareaName: string;
+  /**
+   * @see {@link https://www.tiny.cloud/docs/tinymce/7/react-ref/#tinymcescriptsrc React Tech Ref - tinymceScriptSrc}
+   * @description The URL of the TinyMCE script to lazy load.
+   */
   tinymceScriptSrc: string | string[] | ScriptItem[];
+  /**
+   * @see {@link https://www.tiny.cloud/docs/tinymce/7/react-ref/#rollback React Tech Ref - rollback}
+   * @description The number of milliseconds to wait before reverting to the previous value when the editor's content changes.
+   */
   rollback: number | false;
+  /**
+   * @see {@link https://www.tiny.cloud/docs/tinymce/7/react-ref/#scriptloading React Tech Ref - scriptLoading}
+   * @description Options for how the TinyMCE script should be loaded.
+   * @property async Whether the script should be loaded with the `async` attribute.
+   * @property defer Whether the script should be loaded with the `defer` attribute.
+   * @property delay The number of milliseconds to wait before loading the script.
+   */
   scriptLoading: {
     async?: boolean;
     defer?: boolean;
     delay?: number;
   };
+  /**
+   * @see {@link https://www.tiny.cloud/docs/tinymce/7/react-ref/#licenseKey React Tech Ref - licenseKey}
+   * @description Tiny Cloud License Key for when self-hosting TinyMCE.
+   */
   licenseKey: string;
 }
 
 export interface IAllProps extends Partial<IProps>, Partial<IEvents> { }
 
 /**
- * @see {@link https://www.tiny.cloud/docs/tinymce/7/react-ref/} for the TinyMCE React Technical Reference
+ * @see {@link https://www.tiny.cloud/docs/tinymce/7/react-ref/ TinyMCE React Technical Reference}
  */
 export class Editor extends React.Component<IAllProps> {
   public static propTypes: IEditorPropTypes = EditorPropTypes;
@@ -50,7 +147,7 @@ export class Editor extends React.Component<IAllProps> {
   public editor?: TinyMCEEditor;
 
   private id: string;
-  private elementRef: React.RefObject<HTMLElement>;
+  private elementRef: React.RefObject<HTMLElement | null>;
   private inline: boolean;
   private currentContent?: string;
   private boundHandlers: Record<string, (event: EditorEvent<unknown>) => unknown>;
@@ -60,7 +157,7 @@ export class Editor extends React.Component<IAllProps> {
   public constructor(props: Partial<IAllProps>) {
     super(props);
     this.id = this.props.id || uuid('tiny-react');
-    this.elementRef = React.createRef<HTMLElement>();
+    this.elementRef = React.createRef<HTMLElement | null>();
     this.inline = this.props.inline ?? this.props.init?.inline ?? false;
     this.boundHandlers = {};
   }
@@ -95,7 +192,7 @@ export class Editor extends React.Component<IAllProps> {
                 // getBookmark throws exceptions when the editor has not been focused
                 // possibly only in inline mode but I'm not taking chances
                 cursor = localEditor.selection.getBookmark(3);
-              } catch (e) { /* ignore */ }
+              } catch (_e) { /* ignore */ }
             }
             const valueCursor = this.valueCursor;
             localEditor.setContent(this.props.value as string);
@@ -106,7 +203,7 @@ export class Editor extends React.Component<IAllProps> {
                     localEditor.selection.moveToBookmark(bookmark);
                     this.valueCursor = bookmark;
                     break;
-                  } catch (e) { /* ignore */ }
+                  } catch (_e) { /* ignore */ }
                 }
               }
             }
@@ -146,7 +243,7 @@ export class Editor extends React.Component<IAllProps> {
   public componentWillUnmount() {
     const editor = this.editor;
     if (editor) {
-      editor.off(this.changeEvents(), this.handleEditorChange);
+      editor.off(changeEvents, this.handleEditorChange);
       editor.off(this.beforeInputEvent(), this.handleBeforeInput);
       editor.off('keypress', this.handleEditorChangeSpecial);
       editor.off('keydown', this.handleBeforeInputSpecial);
@@ -162,14 +259,6 @@ export class Editor extends React.Component<IAllProps> {
 
   public render() {
     return this.inline ? this.renderInline() : this.renderIframe();
-  }
-
-  private changeEvents() {
-    const isIE = getTinymce(this.view)?.Env?.browser?.isIE();
-    return (isIE
-      ? 'change keyup compositionend setcontent CommentChange'
-      : 'change input compositionend setcontent CommentChange'
-    );
   }
 
   private beforeInputEvent() {
@@ -238,13 +327,13 @@ export class Editor extends React.Component<IAllProps> {
       const wasControlled = isValueControlled(prevProps);
       const nowControlled = isValueControlled(this.props);
       if (!wasControlled && nowControlled) {
-        this.editor.on(this.changeEvents(), this.handleEditorChange);
+        this.editor.on(changeEvents, this.handleEditorChange);
         this.editor.on(this.beforeInputEvent(), this.handleBeforeInput);
         this.editor.on('keydown', this.handleBeforeInputSpecial);
         this.editor.on('keyup', this.handleEditorChangeSpecial);
         this.editor.on('NewBlock', this.handleEditorChange);
       } else if (wasControlled && !nowControlled) {
-        this.editor.off(this.changeEvents(), this.handleEditorChange);
+        this.editor.off(changeEvents, this.handleEditorChange);
         this.editor.off(this.beforeInputEvent(), this.handleBeforeInput);
         this.editor.off('keydown', this.handleBeforeInputSpecial);
         this.editor.off('keyup', this.handleEditorChangeSpecial);
@@ -264,7 +353,7 @@ export class Editor extends React.Component<IAllProps> {
         if (this.valueCursor && (!this.inline || editor.hasFocus())) {
           try {
             editor.selection.moveToBookmark(this.valueCursor);
-          } catch (e) { /* ignore */ }
+          } catch (_e) { /* ignore */ }
         }
       });
     }
@@ -278,7 +367,7 @@ export class Editor extends React.Component<IAllProps> {
           // getBookmark throws exceptions when the editor has not been focused
           // possibly only in inline mode but I'm not taking chances
           this.valueCursor = this.editor.selection.getBookmark(3);
-        } catch (e) { /* ignore */ }
+        } catch (_e) { /* ignore */ }
       }
     }
   };
@@ -346,7 +435,7 @@ export class Editor extends React.Component<IAllProps> {
     }
 
     const finalInit: EditorOptions = {
-      ...this.props.init,
+      ...this.props.init as Omit<InitOptions, OmittedInitProps>,
       selector: undefined,
       target,
       readonly: this.props.disabled,
@@ -401,6 +490,7 @@ export class Editor extends React.Component<IAllProps> {
       target.value = this.getInitialValue();
     }
 
+    // eslint-disable-next-line @typescript-eslint/no-floating-promises
     tinymce.init(finalInit);
   };
 }